Application-wide refactoring

+++ src/main/java/com/mohsen/springtunes/dao/ArtistDAO.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ArtistDAO extends JpaRepository<Artist, Integer> {
+public interface ArtistDAO extends JpaRepository<Artist, Long> {
 }
